Very first version, mimic the $RANDOM behavior
Creating a gordm package with a single function just to return a random
integer.

The gordm package needs to expose the Seed so we can test it in a
deterministic way.

In this commit I'm also adding code for the CLI offering a way for the
users to interact with the functionalities of the gordm package.