Add --web option
Here I'm not adding another test for the CLI to actually call the
program with '-web', as that would be like a repetition of what was made
in `gordm_test.go`.

What I'm doing is adding a test to check if the '--help' message shows
information about the '-web' flag.

By the way, in this commit we also have an example of how to use a
RegExp with a pattern able to match a string with multiple lines.